Add ability to change Lake Hylia's water level after beating Morpha
There's a gossip stone near the warp pad that you can hit to change the water level. It only works if you've beaten morpha. For non-ER, this stone will only exist as adult, and the child water level is always raised and cannot be controlled. For ER, the stone exists for child and adult, and both ages can control the water level

Also while I was going through the gossip stone asm, I make the stone skip the initial "Responding to your mask..." text

The logic is not yet updated to account for this change