Fix Tower Escape Sequence in ER
"Skip Tower Escape Sequence" was previously forced enabled in ER because disabling to the trade quest timers broke the collapse sequence, even with the attempted fix of disabling the collapse timers. 

This should now be handled properly so the tower escape sequence is fully compatible with ER, including the timer.